ocr_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pdf2image
pillow
